requirements.txt
1.PYTHON 
2.SELENIUM
3.PANDAS
4.NUMPY
5.FIREFOX
6.VS_STUDIO_CODE
7.MYSQL